+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Deprecated
 
+- 決済承認アクションのresult.paymentMethodを非推奨化
+
 ### Removed
 
 ### Fixed